Refactor arithmetic functions to use smaller integer types for improved performance

+++ scalar/src/lib.rs
@@ -1,12 +1,12 @@
-pub fn sum(a: i16, b: i16) -> i16 {
+pub fn sum(a: u8, b: u8) -> u8 {
     a + b
 }
 
-pub fn diff(a: i32, b: i32) -> i32 {
+pub fn diff(a: i16, b: i16) -> i16 {
     a - b
 }
 
-pub fn pro(a: i32, b: i32) -> i32 {
+pub fn pro(a: i8, b: i8) -> i8 {
     a * b
 }
 
@@ -14,7 +14,7 @@ pub fn quo(a: f32, b: f32) -> f32 {
     a / b
 }
 
-pub fn rem(a: f64, b: f64) -> f64 {
+pub fn rem(a: f32, b: f32) -> f32 {
     a % b
 }
